# pqarrow/builder: optimize Parquet to Arrow conversion