--- a/pqarrow/builder/recordbuilder.go
+++ b/pqarrow/builder/recordbuilder.go
@@ -9,6 +9,19 @@ import (
 	"github.com/apache/arrow/go/v8/arrow/memory"
 )
 
+func NewBuilder(mem memory.Allocator, t arrow.DataType) ColumnBuilder {
+	switch t := t.(type) {
+	case *arrow.BinaryType:
+		return NewOptBinaryBuilder(arrow.BinaryTypes.Binary)
+	case *arrow.Int64Type:
+		return NewOptInt64Builder(arrow.PrimitiveTypes.Int64)
+	case *arrow.ListType:
+		return NewListBuilder(mem, t.Elem())
+	default:
+		return array.NewBuilder(mem, t)
+	}
+}
+
 // The code in this file is based heavily on Apache arrow's array.RecordBuilder,
 // with some modifications to use our own optimized record builders. Ideally, we
 // would eventually merge this upstream.
@@ -32,12 +45,7 @@ func NewRecordBuilder(mem memory.Allocator, schema *arrow.Schema) *RecordBuilder
 	}
 
 	for i, f := range schema.Fields() {
-		switch f.Type.(type) {
-		case *arrow.BinaryType:
-			b.fields[i] = NewOptBinaryBuilder(arrow.BinaryTypes.Binary)
-		default:
-			b.fields[i] = array.NewBuilder(b.mem, f.Type)
-		}
+		b.fields[i] = NewBuilder(mem, f.Type)
 	}
 
 	return b